2 components have been created, this is dependant on whether the projects object contains a key of hostedLink. If it does not, then it will render the component with the component with the guthuyb button sized bigger and centered, if it does it renders the div with the hosted link and the github link rendered in a grid with 2 columns
